Functional Icons Container
Updated links to each icon- updated names of components. Moved files and recreated accurate file paths to render on browser. Icons container is now functional. Minimum css styling added. Can update if needed.
