--- client/src/components/Icons-Components/All-IconsContainer.css
@@ -1,5 +1,0 @@
-
-.icons-container {
-    display: flex;
-    align-items: center;
-}
