+++ client/src/components/Icons/GoogleDocIconSvg.jsx
@@ -6,7 +6,7 @@ const GoogleDocsLogosvg = () => {
     return (
        < a 
        className="google-doc-project-brief-icon" 
-        href="https://github.com/Gayle-Thompson-Igwebike/GOOD-PR-v1">
+        href="https://docs.google.com/document/d/1O-FOKo0n4gX96g5b_wm65zurs463fkdbTcGdbmxx79Q/edit">
        <img src={GoogleDocIcon} alt="google-doc-project-brief-icon-img"  /> 
        </a>
     )
