+++ client/src/components/Icons/SlackIconSvg.jsx
@@ -7,8 +7,8 @@ const SlackIconSvg = () => {
     return (
        < a 
        className="slack-icon" 
-        href="https://github.com/Gayle-Thompson-Igwebike/GOOD-PR-v1">
-       <img src={SlackIcon} alt="slack icon"/> 
+        href="https://codeyourfuture.slack.com/archives/C057SMMPG8Y">
+       <img src={SlackIcon} alt="slack icon"/> <style> height: 200px</style>
        </a>
     )
